moved internal.core.deployment package to internal.deployment
