+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/UUFServer.java
@@ -33,8 +33,8 @@ import org.wso2.carbon.kernel.startupresolver.RequiredCapabilityListener;
 import org.wso2.carbon.uuf.api.Server;
 import org.wso2.carbon.uuf.core.App;
 import org.wso2.carbon.uuf.exception.UUFException;
-import org.wso2.carbon.uuf.internal.core.deployment.AppDeployer;
-import org.wso2.carbon.uuf.internal.core.deployment.DeploymentNotifier;
+import org.wso2.carbon.uuf.internal.deployment.AppDeployer;
+import org.wso2.carbon.uuf.internal.deployment.DeploymentNotifier;
 import org.wso2.carbon.uuf.internal.io.ArtifactAppDeployer;
 import org.wso2.carbon.uuf.spi.HttpConnector;
 import org.wso2.carbon.uuf.spi.HttpRequest;
