+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/deployment/parser/DependencyTreeParser.java
@@ -16,34 +16,33 @@
  * under the License.
  */
 
-package org.wso2.carbon.uuf.internal.core.deployment.parser;
+package org.wso2.carbon.uuf.internal.deployment.parser;
 
+import org.wso2.carbon.uuf.api.config.DependencyNode;
 import org.wso2.carbon.uuf.api.reference.FileReference;
 import org.wso2.carbon.uuf.exception.MalformedConfigurationException;
 import org.yaml.snakeyaml.Yaml;
 
-import java.util.Map;
-
 /**
- * Parser for configuration file in an UUF App.
+ * Parser for dependency tree file in an UUF App.
  *
  * @since 1.0.0
  */
-public class ConfigurationParser {
+public class DependencyTreeParser {
 
     /**
-     * Parses the specified config YAML file.
+     * Parses the specified dependency tree YAML file.
      *
-     * @param configFile reference to the config YAML file
-     * @return configuration as a {@link Map}
-     * @exception MalformedConfigurationException if cannot parse the specified configuration file
+     * @param dependencyTreeFile reference to the dependency tree YAML file
+     * @return root node of the dependecy tree
+     * @exception MalformedConfigurationException if cannot parse specified dependency tree file
      */
-    public static Map<?, ?> parse(FileReference configFile) {
+    public static DependencyNode parse(FileReference dependencyTreeFile) {
         try {
-            return new Yaml().loadAs(configFile.getContent(), Map.class);
+            return new Yaml().loadAs(dependencyTreeFile.getContent(), DependencyNode.class);
         } catch (Exception e) {
             throw new MalformedConfigurationException(
-                    "Cannot parse configuration file '" + configFile.getAbsolutePath() + "'.", e);
+                    "Cannot parse dependency tree file '" + dependencyTreeFile.getAbsolutePath() + "'.", e);
         }
     }
 }
