+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/io/ArtifactAppDeployer.java
@@ -25,9 +25,9 @@ import org.wso2.carbon.uuf.core.App;
 import org.wso2.carbon.uuf.exception.FileOperationException;
 import org.wso2.carbon.uuf.exception.UUFException;
 import org.wso2.carbon.uuf.internal.UUFServer;
-import org.wso2.carbon.uuf.internal.core.deployment.AppCreator;
-import org.wso2.carbon.uuf.internal.core.deployment.AppDeployer;
-import org.wso2.carbon.uuf.internal.core.deployment.ClassLoaderProvider;
+import org.wso2.carbon.uuf.internal.deployment.AppCreator;
+import org.wso2.carbon.uuf.internal.deployment.AppDeployer;
+import org.wso2.carbon.uuf.internal.deployment.ClassLoaderProvider;
 import org.wso2.carbon.uuf.internal.io.util.ZipArtifactHandler;
 import org.wso2.carbon.uuf.internal.util.NameUtils;
 import org.wso2.carbon.uuf.spi.RenderableCreator;
