+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/io/BundleClassLoaderProvider.java
@@ -31,7 +31,7 @@ import org.wso2.carbon.uuf.api.reference.ComponentReference;
 import org.wso2.carbon.uuf.api.reference.FileReference;
 import org.wso2.carbon.uuf.exception.FileOperationException;
 import org.wso2.carbon.uuf.exception.UUFException;
-import org.wso2.carbon.uuf.internal.core.deployment.ClassLoaderProvider;
+import org.wso2.carbon.uuf.internal.deployment.ClassLoaderProvider;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
